# cpu: riscv: gemm: add f32 gemm SIMD optimization with RISC-V V Extension