--- a/src/cpu/gemm/gemm.cpp
+++ b/src/cpu/gemm/gemm.cpp
@@ -48,7 +48,7 @@ using namespace dnnl::impl::cpu::ppc64;
 #elif DNNL_S390X
 #include "cpu/s390x/gemm.h"
 #elif DNNL_RV64
-#if DNNL_RISCV_USE_RVV_INTRINSICS
+#if defined(DNNL_RISCV_USE_RVV_INTRINSICS)
 #include "cpu/rv64/gemm/rvv_gemm_f32.hpp"
 using namespace dnnl::impl::cpu::rv64;
 #endif
@@ -150,7 +150,7 @@ dnnl_status_t extended_sgemm(const char *transa, const char *transb,
     }
 #endif
 
-#if DNNL_RV64 && DNNL_RISCV_USE_RVV_INTRINSICS
+#if DNNL_RV64 && defined(DNNL_RISCV_USE_RVV_INTRINSICS)
     return rvv_gemm_f32(
             transa, transb, M, N, K, alpha, A, lda, B, ldb, beta, C, ldc, bias);
 #endif