# Description

This PR introduces a SIMD-optimized f32 GEMM kernel for the RISC-V 64-bit architecture, leveraging the RISC-V Vector (V) Extension.

The existing generic GEMM implementation in oneDNN (`ref_gemm`) relies on `PRAGMA_OMP_SIMD` for auto-vectorization. However, current RISC-V toolchains do not effectively vectorize this code, leading to suboptimal performance on RV64 platforms. This commit addresses this critical performance gap by providing a manually optimized SIMD implementation, `rvv_gemm_f32`, which is dispatched from `extended_sgemm`.

This optimization is integrated directly into the low-level `extended_sgemm` function. This provides two major benefits:
1.  It accelerates direct `sgemm` calls across the library.
2.  It boosts the performance of high-level primitives that rely on this function. The example is `gemm_convolution_fwd_t`, one of the most efficient forward convolution implementations in oneDNN.

By targeting the foundational GEMM routine, this PR delivers performance gains for models like Convolutional Neural Networks (CNNs).

The current implementation only focuses on the `f32` data type.

# Checklist

## General

- [√] Do all unit and benchdnn tests (`make test` and `make test_benchdnn_*`) pass locally for each commit?
- [√] Have you formatted the code using clang-format?

## Performance improvements

- [√] Have you submitted performance data that demonstrates performance improvements?  

All performance data was measured on a Spacemit X60 development board (VLEN=256). The performance baseline is the default `ref_gemm` implementation in the oneDNN main branch. Below are some example performance benchmarks for different problem sizes.

### Matmul Primitives Performance

| Benchmark Case                                                | Baseline (`ref_gemm`) | This PR (`rvv_gemm`) | Speedup |
| :------------------------------------------------------------ | :-------------------- | :------------------- | :------ |
| `--matmul --stag=acb --wtag=cab 2x457x3888:2x3888x2888`        | 11.44s                | 3.28s                | 3.49x |
| `--matmul --bia-dt=f32 1024x1024:1024x1024`                    | 3.34s                 | 2.53s                | 1.32x |
| `--matmul --stag=acb --wtag=abc 2x551x1276:2x1276x58`          | 0.09s                 | 0.04s                | ~2.25x |

### Convolution Primitives Performance (via GEMM)

This demonstrates the advantage of optimizing the `extended_sgemm`.

| Benchmark Case                                                                                            | Baseline (`gemm_conv`) | This PR (`gemm_conv` with `rvv_gemm`) | Speedup |
| :-------------------------------------------------------------------------------------------------------- | :-------------------- | :------------------ | :------ |
| `--conv mb1ic3id116ih132iw132oc32od114oh130ow130kd3kh3kw3pd0ph0pw0n"3d_unet:conv_1"` | 2.09s                 | 1.55s               | 1.35x |

### New features

- [  ] Have you published an RFC for the new feature?
- [  ] Was the RFC approved?
- [  ] Have you added relevant tests?

### Bug fixes

- [  ] Have you included information on how to reproduce the issue (either in a github issue or in this PR)?
- [  ] Have you added relevant regression tests?

## [RFC](https://github.com/uxlfoundation/oneDNN/tree/rfcs) PR

- [  ] Does RFC document follow the [template](https://github.com/uxlfoundation/oneDNN/blob/rfcs/rfcs/template.md#onednn-design-document-rfc)?
- [  ] Have you added a link to the rendered document?