@xiazhuozhao 2025-08-28
Hello everyone,

Thank you for your patience. I've updated the PR with a new commit to address the performance instability issues observed in the initial submission. Here's a summary of the changes and their impact:

### Key Updates in the New Commit:
1.  **Inner Block Size Adjustment**: The F32 GEMM kernel's inner block size has been set to 8 to better align with the 256-bit vector length (VLEN), improving vector unit utilization. For future work, optimizations for various VLENs can be implemented using templates, which aligns with the variable-length design of the RISC-V V extension.  
2.  **Fallback for TransA Case**: The TransA (A^T * B) case has been fallbacked to the `ref_gemm_f32` implementation. The previous RVV-specific optimization for this path was inefficient and caused regressions.  
3.  **Improved Register Utilization**: Manual loop unrolling for the n=4 case has been implemented. This avoids spilling `vfloat32m1_t` variables to the stack.  

### Performance Impact:
This optimization has been tested, showing performance gains across all benchmarks. It yields significant speedups for Matmul (averaging 4.00x) and Deconvolution (averaging 5.16x) primitives, while the overall impact on convolution is generally positive.  

#### Detailed Performance Data:

**Matmul Performance:**
[mm.csv](https://static.xiazhuozhao.com/temp/mm.csv)

**Convolution Performance:**
[conv.csv](https://static.xiazhuozhao.com/temp/conv.csv)

**Deconvolution Performance:**
[deconv.csv](https://static.xiazhuozhao.com/temp/deconv.csv)

This GEMM optimization effectively enhances the performance of the default convolution and deconvolution implementations in RVV, as they directly invoke GEMM rather than Matmul. Consequently, this submission does not conflict with any existing matrix multiplication optimizations. We kindly request that you consider this PR for approval.