@xiazhuozhao 2025-08-28
Hi @spalicki ,
Sorry to bother you directly. I'm hoping to get a review for my PR #3785 .
I'm reaching out to you specifically because I noticed that you reviewed a similar GEMM optimization PR #3156  for the ppc64 architecture in the past. Your detailed feedback in that review helps me a lot. If you have a moment, would you be willing to take a look? Any guidance or feedback would be greatly appreciated.
Thank you so much for your time and consideration!