@xiazhuozhao 2025-09-04
Hi @dzarukin,

I have now implemented all the changes from your feedback, and I'm happy to report that all local tests are passing successfully.

During the testing process, I noticed that the excellent PR #3784 was successfully merged. In light of its contributions, the matmul performance of this PR will now align with the improvements introduced there, GEMM-based matmul will no longer be used by default on RVV devices.

However, this PR continues to provide value by enhancing the performance of extended_sgemm on the RVV platform, this accelerates primitives like convolution and deconvolution, as well as any other operation that calls the extended_sgemm function, for instance, ref_rnn.

Furthermore, the gemm kernel introduced here is designed with greater long-term optimization potential. The current implementation is a balanced trade-off for VLEN=256, and in the future, this framework will allow us to easily tune key parameters like m and n to deliver even better performance for variable vector lengths (such as 128, 512, etc.) across different hardware.

Thank you once again for your invaluable time and guidance!