@Ralap-Jee 2025-09-15
> in my experience, it might have been translated into AVX instructions, making its performance more akin to an X86 processor.

Agreed. QEMU simulation results should only be used to verify functional correctness, not for performance evaluation.

> Additionally, it's important to note that the current rvv_gemm is a compromise optimized for VLEN=256, whereas the SG2044's VLEN is 512, meaning half of the theoretical performance has yet to be unlocked.

It is confusing that which line of your code shows specific optimization of VLEN=256? In my understanding, `vl` is runtime-determined by `vsetvl` with `avl` and `VLEN`, which is suitable for machines with different VLENs. How does performance differ?