@zhangjian29 2025-09-19
> It's important to note that the previous performance data was collected on a low-performing 8-core Spacemit(R) X60 development board. Its limited memory bandwidth prevented it from accurately showcasing the optimization's full effect.

Hi, I am facing a similar situation. May I kindly ask: is the memory bandwidth truly the key bottleneck causing the poor performance on the Spacemit X60 board?

To clarify, does this imply that programs using RVV intrinsics require more memory bandwidth to leverage their strengths effectively compared to scalar implementations? Interesting! Do you got any research papers to have it proved?